Small refactor and PHPDoc blocks on nodes
