+++ src/JsPhpize/Nodes/ArrayBase.php
@@ -2,6 +2,11 @@
 
 namespace JsPhpize\Nodes;
 
+/**
+ * Class ArrayBase.
+ *
+ * @property-read array $data array elements
+ */
 abstract class ArrayBase extends Value
 {
     protected $data = [];
